requirements.txt
Flask==3.0.0
requests==2.31.0
pandas==2.1.1
python-dotenv==1.0.0
plotly==5.17.0
gunicorn==21.2.0
